+++ packages/tizen_rpc_port/lib/src/stub_base.dart
@@ -9,16 +9,15 @@ import 'package:ffi/ffi.dart';
 import 'package:flutter/services.dart';
 import 'package:tizen_interop/6.5/tizen.dart';
 
-import 'package:tizen_log/tizen_log.dart';
 import 'parcel.dart';
 import 'port.dart';
-import 'rpc_port_method_channel.dart';
+import 'proxy_base.dart' show OnError;
 
 export 'package:meta/meta.dart' show nonVirtual, visibleForOverriding;
 
-/// The abstract class for creating a stub class for RPC.
+/// The base class for stub objects used by RPC servers.
 abstract class StubBase {
-  /// The constructor for this class.
+  /// Creates a [StubBase] instance with the specified port name.
   StubBase(this.portName) {
     _handle = using((Arena arena) {
       final Pointer<rpc_port_stub_h> pStub = arena();
@@ -30,7 +29,6 @@ abstract class StubBase {
           message: tizen.get_error_message(ret).toDartString(),
         );
       }
-
       return pStub.value;
     });
 
@@ -39,23 +37,23 @@ abstract class StubBase {
 
   late final rpc_port_stub_h _handle;
 
-  static const String _logTag = 'RPC_PORT';
-
-  static final MethodChannelRpcPort _methodChannel =
-      MethodChannelRpcPort.instance;
-
   final Finalizer<StubBase> _finalizer = Finalizer<StubBase>((StubBase stub) {
     stub._streamSubscription?.cancel();
     tizen.rpc_port_stub_destroy(stub._handle);
   });
 
-  /// The port name of the connection with the proxy.
+  /// A port name to use when listening for connections.
   final String portName;
 
-  /// ignore: cancel_subscriptions
+  static const EventChannel _eventChannel = EventChannel('tizen/rpc_port_stub');
+
+  // ignore: cancel_subscriptions
   StreamSubscription<dynamic>? _streamSubscription;
 
-  /// Sets a trusted proxy to the stub.
+  /// Sets whether this stub should only allow trusted connections.
+  ///
+  /// If [trusted] is true, this stub only allows connections from trusted
+  /// applications signed with the same certificate.
   void setTrusted(bool trusted) {
     final int ret = tizen.rpc_port_stub_set_trusted(_handle, trusted);
     if (ret != 0) {
@@ -66,7 +64,7 @@ abstract class StubBase {
     }
   }
 
-  /// Adds a privilege which is required when the proxy connects this stub.
+  /// Adds a privilege required by client applications to connect to this stub.
   void addPrivilege(String privilege) {
     using((Arena arena) {
       final Pointer<Char> pPrivilege = privilege.toNativeChar(allocator: arena);
@@ -80,68 +78,59 @@ abstract class StubBase {
     });
   }
 
-  /// Listens to the requests for connections.
-  Future<void> listen() async {
+  /// Starts listening to connection requests from clients.
+  Future<void> listen({OnError? onError}) async {
     if (_streamSubscription != null) {
       return;
     }
 
-    final Stream<dynamic> stream =
-        await _methodChannel.stubListen(_handle.address);
-    _streamSubscription = stream.listen((dynamic event) async {
-      if (event is Map) {
-        final Map<dynamic, dynamic> map = event;
-        final int handle = map['handle'] as int;
-        if (handle != _handle.address) {
-          return;
-        }
-
-        final String eventName = map['event'] as String;
-        final String sender = map['sender'] as String;
-        final String instance = map['instance'] as String;
-
-        Log.info(
-            _logTag, 'event: $eventName, sender: $sender, instance: $instance');
-        if (eventName == 'connected') {
-          await onConnectedEvent(sender, instance);
-        } else if (eventName == 'disconnected') {
-          await onDisconnectedEvent(sender, instance);
-        } else if (eventName == 'received') {
-          final Uint8List rawData = map['rawData'] as Uint8List;
-          final Parcel parcel = Parcel.fromRaw(rawData);
-          await onReceivedEvent(sender, instance, parcel);
-        } else {
-          Log.error(_logTag, 'Unknown event: $eventName');
-        }
+    final Stream<dynamic> stream = _eventChannel
+        .receiveBroadcastStream(<String, Object>{'handle': _handle.address});
+    _streamSubscription = stream.listen((dynamic map) async {
+      final int handle = map['handle'] as int;
+      if (handle != _handle.address) {
+        return;
       }
-    });
+      final String event = map['event'] as String;
+      final String sender = map['sender'] as String;
+      final String instance = map['instance'] as String;
+      if (event == 'connected') {
+        await onConnectedEvent(sender, instance);
+      } else if (event == 'disconnected') {
+        await onDisconnectedEvent(sender, instance);
+      } else if (event == 'received') {
+        final Uint8List rawData = map['rawData'] as Uint8List;
+        final Parcel parcel = Parcel.fromRaw(rawData);
+        await onReceivedEvent(sender, instance, parcel);
+      } else {
+        print('Unknown event: $event');
+      }
+    }, onError: onError);
   }
 
-  /// Gets a port connected with proxy that has instance id.
+  /// Gets a [Port] associated with a specific client [instance].
   Port getPort(String instance, PortType portType) {
     return using((Arena arena) {
       final Pointer<Char> pInstance = instance.toNativeChar();
       final Pointer<rpc_port_h> pPort = arena();
       final int ret = tizen.rpc_port_stub_get_port(
           _handle, portType.index, pInstance, pPort);
-
       if (ret != 0) {
         throw PlatformException(
           code: ret.toString(),
           message: tizen.get_error_message(ret).toDartString(),
         );
       }
-
       return Port.fromNativeHandle(pPort.value);
     });
   }
 
-  /// The abstract method for receiving connected event.
+  /// Called when a connection is established by a client.
   Future<void> onConnectedEvent(String sender, String instance);
 
-  /// The abstract method for receiving disconnected event.
+  /// Called when a connection is closed by a client.
   Future<void> onDisconnectedEvent(String sender, String instance);
 
-  /// The abstract method called when the stub receives data from proxy.
+  /// Called when data are received from a client.
   Future<void> onReceivedEvent(String sender, String instance, Parcel parcel);
 }