+++ packages/tizen_rpc_port/tizen/src/rpc_port_stub.h
@@ -2,34 +2,27 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef RPC_PORT_STUB_H
-#define RPC_PORT_STUB_H
+#ifndef FLUTTER_PLUGIN_RPC_PORT_STUB_H_
+#define FLUTTER_PLUGIN_RPC_PORT_STUB_H_
 
-#include <flutter/event_channel.h>
-#include <flutter/standard_method_codec.h>
-#include <rpc-port-parcel.h>
+#include <flutter/encodable_value.h>
+#include <flutter/event_sink.h>
 #include <rpc-port.h>
 
-#include <map>
 #include <memory>
-#include <set>
 #include <string>
-#include <vector>
 
-#include "common.h"
+#include "rpc_port_result.h"
 
 namespace tizen {
 
-typedef std::unique_ptr<flutter::EventSink<flutter::EncodableValue>> EventSink;
+typedef flutter::EventSink<flutter::EncodableValue> FlEventSink;
 
 class RpcPortStubManager {
  public:
-  static void Init(EventSink sync);
+  static void Init(std::unique_ptr<FlEventSink> event_sink);
 
   static RpcPortResult Listen(rpc_port_stub_h handle);
-  static RpcPortResult AddPrivilege(rpc_port_stub_h handle,
-                                    const std::string& privilege);
-  static RpcPortResult SetTrusted(rpc_port_stub_h handle, const bool trusted);
 
  private:
   static void OnConnectedEvent(const char* sender, const char* instance,
@@ -39,10 +32,9 @@ class RpcPortStubManager {
   static int OnReceivedEvent(const char* sender, const char* instance,
                              rpc_port_h port, void* user_data);
 
- private:
-  static EventSink event_sink_;
+  static std::unique_ptr<FlEventSink> event_sink_;
 };
 
 }  // namespace tizen
 
-#endif  // RPC_PORT_STUB_H
+#endif  // FLUTTER_PLUGIN_RPC_PORT_STUB_H_