+++ packages/tizen_rpc_port/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.1
+
+* Clean up the API documentation.
+* Add optional `OnError` callback arguments to `ProxyBase.connect` and `StubBase.listen`.
+* Refactor the code and fix bugs.
+* Revise the example apps.
+
 ## 0.1.0
 
 * Initial release.